// src/utils/enums/user-category.enum.ts
export enum AppUserCategory {
  MILITARY = 'MILITARY',
  MILITARY_FAMILY = 'MILITARY_FAMILY',
  CIVILIAN = 'CIVILIAN',
  VETERANS = 'VETERANS',
  MINDEF_EMPLOY = 'MINDEF EMPLOY',
}
export enum UserCategory {
  ADMIN = 'ADMIN',
  MILITARY = 'MILITARY',
  MIN_DEF = 'MINDEF',
  ARMY = 'ARMY',
  NAVY = 'NAVY',
  AIR_FORCE = 'AIR FORCE',
  CHIEF_ARMED_FORCES_PAT = 'CHIEF ARMED FORCES (PAT)',
  JHEV = 'JHEV',
}